backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
sqlalchemy==2.0.36
psycopg2-binary==2.9.10
alembic==1.14.0
pydantic==2.10.4
pydantic-settings==2.7.0
python-jose[cryptography]==3.3.0
bcrypt==4.2.1
python-multipart==0.0.20
email-validator==2.2.0
scikit-learn==1.6.0
numpy==2.2.1
pandas==2.2.3
joblib==1.4.2
weasyprint==63.1
jinja2==3.1.5
openpyxl==3.1.5
redis==5.2.1
celery==5.4.0
python-dotenv==1.0.1
pytest==8.3.4
httpx==0.28.1
